Refactored the code in components
